Increase wait-init-complete timeout
When a host starts, the systemd service xapi-wait-init-complete waiting
on the creation of the xapi init cookie file may fail on timeout for a
matter of seconds. This patch adds 1 minute (300 seconds total) to the
timeout passed to the script xapi-wait-init-complete.

The patch is merged upstream (commit 801dd96cd) and has been modified to
apply on top of xapi-1.249.38.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>