Enable experimental components and http endpoint
To enable SMAPI traces we need to enable experimental components. By
default SMAPI is experimental and it is always filtered by XAPI. Note
that currently SMAPI is the only experimental components.
This patch also enables to use an HTTP endpoint to sent traces. This
will allow to setup a jaeger server and collect traces.

Signed-off-by: Guillaume <guillaume.thouvenin@vates.tech>