Remove dependencies on unused components
We were not using hcp_nss before, and now that it's been renamed to
nss-override-id, nothing changed. But we are evaluating possible use
cases, so keep it as a comment (also for easier rebasing in the future).

We are not using the sysprep feature (which is disabled in xapi by
default), so comment out requirements for it as we don't want to install
packages that will go unused.

Also drop dependency on xapi-ssh-monitor. It is currently proprietary and will
only provide a daemon that handles SSH access in xapi's absence - we are not
exposing the SSH automode feature for production use at the moment, so it's
not critical.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>